Use ghcr.io for registry in acceptance tests also
I was still getting timeouts, and Cursor pointed out that I'd hit a
rate limit on docker.io.

This change might be fixing a temporary problem, but I can't see any
harm in fetching this from ghcr.io instead of docker.io. And I
believe the ghcr.io has less rate limits in general. (Supposedly
docker.io rate limits after 100 pulls in 6 hours.)

Note that we're also updating to a newer version here. The older
stable version is 2.8.3, but that's not available in this repo.

Ref: https://issues.redhat.com/browse/EC-1387